implement search and contains and refactor project
