+++ Didukh_Lopatetskiy_FI-13_14/.gitignore
@@ -1,0 +1,2 @@
+/build/
+/.vscode/
