+++ Didukh_Lopatetskiy_FI-13_14/tests/CMakeLists.txt
@@ -1,14 +1,16 @@
 find_package(GTest REQUIRED)
 
-add_executable(tests
-    tests.cpp
+add_executable(runTests
+    invertedIndexTests.cpp
+    parserTests.cpp
+    runTests.cpp
 )
 
-target_link_libraries(tests
+target_link_libraries(runTests
     InvertedIndex
     GTest::GTest 
     GTest::Main   
 )
 
 include(GoogleTest)
-gtest_discover_tests(tests)  
+gtest_discover_tests(runTests)
