+++ Didukh_Lopatetskiy_FI-13_14/tests/invertedIndexTests.cpp
@@ -1,0 +1,11 @@
+#include <invertedIndex.hpp>
+#include <gtest/gtest.h>
+
+class InvertedIndexTests : public ::testing::Test {
+protected:
+    Parser parser;
+};
+
+TEST_F(InvertedIndexTests, test) {
+    ASSERT_EQ(1, 1);
+}
