+++ Didukh_Lopatetskiy_FI-13_14/tests/runTests.cpp
@@ -1,0 +1,7 @@
+#include <invertedIndex.hpp>
+#include <gtest/gtest.h>
+
+int main(int argc, char** argv) {
+    ::testing::InitGoogleTest(&argc, argv);
+    return RUN_ALL_TESTS();
+}
